Issue #116 feat : WriteModel에 추가 후 sharedCount를 1 올리는 테스트 진행
