+++ app/src/main/java/com/lateinit/rightweight/data/model/WriteModel.kt
@@ -1,5 +1,7 @@
 package com.lateinit.rightweight.data.model
 
+import com.google.gson.annotations.SerializedName
+import com.lateinit.rightweight.data.remote.model.IntValue
 import com.lateinit.rightweight.data.remote.model.RemoteData
 
 data class WriteRequestBody(
@@ -8,6 +10,7 @@ data class WriteRequestBody(
 
 
 data class WriteModelData(
+    val transform: TransformData? = null,
     val update: UpdateData? = null,
     val delete: String? = null,
 ){
@@ -19,4 +22,14 @@ data class WriteModelData(
 data class UpdateData(
     val name: String,
     val fields: RemoteData
+)
+
+data class TransformData(
+    val document: String,
+    val fieldTransforms: List<FieldTransformsModelData>
+)
+
+data class FieldTransformsModelData(
+    val fieldPath: String,
+    val increment: IntValue
 )
